Added Word List &  'Agenda' for Job Processing
